+++ ios/ios/src/Web Services/CMWebService.m
@@ -325,26 +325,15 @@ NSString * const YAJLErrorKey = @"YAJLErrorKey";
 #pragma mark - Snippet execution
 
 - (void)runSnippet:(NSString *)snippetName withParams:(NSDictionary *)params user:(CMUser *)user successHandler:(CMWebServiceSnippetRunSuccessCallback)successHandler errorHandler:(CMWebServiceSnippetRunFailureCallback)errorHandler {
-    NSString *baseURL = self.apiUrl;
-    NSString *appId = _appIdentifier;
-    NSString *paramStr = @"";
-
+    
+    NSURL *url = [NSURL URLWithString:[NSString stringWithFormat:@"%@/app/%@/run/%@", self.apiUrl, _appIdentifier, snippetName]];
+    
     if (params) {
-        NSMutableArray* queryComponents = [NSMutableArray array];
         for (id key in params) {
-            [queryComponents addObject:[NSString stringWithFormat:@"%@=%@", key, [params objectForKey:key]]];
+            url = [NSURL URLWithString:[[url URLByAppendingAndEncodingQueryString:[NSString stringWithFormat:@"%@=%@", key, [params objectForKey:key]]] absoluteString]];
         }
-
-        paramStr = [queryComponents componentsJoinedByString:@"&"];
     }
-
-    NSMutableString* urlStr = [NSMutableString stringWithFormat:@"%@/app/%@/run/%@", baseURL, appId, snippetName];
-
-    if ([paramStr length]) {
-        [urlStr appendFormat:@"?%@", paramStr];
-    }
-
-    NSURL* url = [NSURL URLWithString:urlStr];
+    
     NSMutableURLRequest* request = [self constructHTTPRequestWithVerb:@"GET" URL:url appSecret:_appSecret binaryData:NO user:user];
     [self executeRequest:request successHandler:^(NSDictionary *results, NSDictionary *errors, NSDictionary *meta, id snippetResult, NSNumber *count, NSDictionary *headers) {
         successHandler(snippetResult, headers);